build: added build-defaults

+++ build.cake
@@ -1,22 +1,10 @@
 #tool "nuget:?package=xunit.runner.console&version=2.4.1"
 #tool "nuget:?package=GitVersion.CommandLine&version=4.0.0"
 
+#l "./build.defaults.cake"
+
 var target = string.IsNullOrEmpty(Argument("target", "Default")) ? "Default" : Argument("target", "Default");
 string nugetFeed = EnvironmentVariable("NUGET_FEED") ?? null;
-
-class Bld
-{
-	public string SrcFolder = "./src";
-	public string DemosFolder = "./src";
-	public string ArtifactsFolder = "./artifacts";
-	public string[] CleanFilters => new []
-	{
-		$"{SrcFolder}/**/bin/",
-		$"{SrcFolder}/**/obj/",
-		$"{DemosFolder}/**/bin/",
-		$"{DemosFolder}/**/obj/",
-	};
-}
 
 var bld = new Bld();
 
@@ -27,15 +15,20 @@ Task("Clean")
 		foreach(var filter in bld.CleanFilters)
 			CleanDirectories(GetDirectories(filter));
 
-		DeleteDirectory(bld.ArtifactsFolder, new DeleteDirectorySettings
-		{
-			Force = true,
-			Recursive = true
-		});
+		if(DirectoryExists(bld.ArtifactsFolder))
+			DeleteDirectory(bld.ArtifactsFolder, new DeleteDirectorySettings
+			{
+				Force = true,
+				Recursive = true
+			});
 	});
 
 Task("Restore")
-	.IsDependentOn("Clean");
+	.IsDependentOn("Clean")
+	.Does(() =>
+	{
+
+	});
 
 Task("Build")
 	.IsDependentOn("Restore");
